requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
qdrant-client==1.9.0
psycopg2-binary==2.9.9
openai==1.3.0
pydantic>=2.5.0
numpy>=1.26.4
sqlalchemy
tqdm
scipy
sentence-transformers